AI-service/requirements.txt
fastapi
uvicorn
pika
python-dotenv
requests
langchain
langchain-openai
langchain-community
chromadb
tiktoken